chore: reorganize components into sub-folders and fix imports
