--- src/components/PopularTrendsSection.tsx
@@ -1,10 +1,0 @@
-
-function PopularTrendsSection() {
-    return (
-        <section className='w-1/3 min-h-screen outline outline-2 outline-black dark:outline-white rounded-md p-1 m-1 shrink'>
-            <h2 className='font-primary-title-roboto'>Trends</h2>
-        </section>
-    )
-}
-
-export default PopularTrendsSection;
